Move alpha tests to separate file.
